requirements.txt
streamlit
tensorflow
numpy
pandas
opencv-python
Pillow
matplotlib
